Add values_into function.
This allows copying a Futhark array directly into an array slice, thus
saving a copy in some cases.

The downside is that smlfut now generates C code, which must be linked
with the final program.  I think it was inevitable that we would reach
this point eventually (especially if we wish to support other SML
compilers), so I am not too hung up about it.